backend code refactor and implementation in user and product done

+++ src/main/java/com/miu/waafinalproject/SecureOnlineAuctionSystem/model/Customer.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int customerID;
+    private Long customerID;
     @OneToOne
     private Users users;
     @OneToMany(mappedBy = "customer")
